+++ src/isomedia/box_code_base.c
@@ -529,11 +529,11 @@ GF_Err cslg_box_write(GF_Box *s, GF_BitStream *bs)
 	e = gf_isom_full_box_write(s, bs);
 	if (e) return e;
 	if (s->version == 0) {
-		gf_bs_write_int(bs, ptr->compositionToDTSShift, 32);
-		gf_bs_write_int(bs, ptr->leastDecodeToDisplayDelta, 32);
-		gf_bs_write_int(bs, ptr->greatestDecodeToDisplayDelta, 32);
-		gf_bs_write_int(bs, ptr->compositionStartTime, 32);
-		gf_bs_write_int(bs, ptr->compositionEndTime, 32);
+		gf_bs_write_int(bs, (s32) ptr->compositionToDTSShift, 32);
+		gf_bs_write_int(bs, (s32) ptr->leastDecodeToDisplayDelta, 32);
+		gf_bs_write_int(bs, (s32) ptr->greatestDecodeToDisplayDelta, 32);
+		gf_bs_write_int(bs, (s32) ptr->compositionStartTime, 32);
+		gf_bs_write_int(bs, (s32) ptr->compositionEndTime, 32);
 		return GF_OK;
 	} else if (s->version == 1) {
 		gf_bs_write_long_int(bs, ptr->compositionToDTSShift, 64);
